+++ src/HRDNS/Types/Stack/FIFO.php
@@ -10,7 +10,7 @@ class FIFO extends Stack
     /**
      * @return integer
      */
-    public function key(): mixed
+    public function key(): int
     {
         return $this->position;
     }